requirements.txt
# Core RAG Dependencies
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
chromadb==0.4.22
sentence-transformers==2.2.2

# Document Processing
pypdf2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
markdown==3.5.1

# Web Framework
streamlit==1.29.0
streamlit-chat==0.1.1

# ML & Embeddings
openai==1.7.2
tiktoken==0.5.2
numpy==1.24.3
faiss-cpu==1.7.4

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
tqdm==4.66.1

# Evaluation
ragas==0.1.0
datasets==2.16.1

# Optional: Advanced features
rank-bm25==0.2.2
transformers==4.36.2
torch==2.1.2
